refactor: API Routes에서 사용하는 상품 더미 데이터를 shared/mock-data/goods.ts로 모듈화
