+++ src/features/main/api/getGoods.mock.ts
@@ -1,5 +1,6 @@
 import { HttpHandler, HttpResponse, http } from "msw";
 import { getGoodsURL } from "./getGoods";
+import { goods } from "@/shared/mock-data/goods";
 
 export const getMockGoods: HttpHandler = http.get(getGoodsURL, () => {
   return HttpResponse.json(GET_MOCK_GOODS.success);
@@ -7,858 +8,7 @@ export const getMockGoods: HttpHandler = http.get(getGoodsURL, () => {
 
 export const GET_MOCK_GOODS = {
   success: {
-    goods: [
-      {
-        id: "1",
-        name: "구름 소파",
-        brand: {
-          id: "1",
-          name: "플러피누들",
-        },
-        price: {
-          originalPrice: "256700",
-          sellingPrice: "154000",
-          discountRate: "40",
-          originalPriceDisplayText: "256,700",
-          sellingPriceDisplayText: "154,000",
-        },
-        imageUrl: "cloud-sofa.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "12:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 9405,
-          reviewAverage: 4.79,
-          reviewCountDisplayText: "9,405",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 87547,
-        },
-      },
-      {
-        id: "2",
-        name: "스파이럴 책장",
-        brand: {
-          id: "2",
-          name: "트위스티코",
-        },
-        price: {
-          originalPrice: "174100",
-          sellingPrice: "155000",
-          discountRate: "11",
-          originalPriceDisplayText: "174,100",
-          sellingPriceDisplayText: "155,000",
-        },
-        imageUrl: "bookshelf.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "12:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 33889,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "33,889",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 66832,
-        },
-      },
-      {
-        id: "3",
-        name: "별빛 펜던트 조명",
-        brand: {
-          id: "3",
-          name: "글리머라이트",
-        },
-        price: {
-          originalPrice: "81200",
-          sellingPrice: "40600",
-          discountRate: "50",
-          originalPriceDisplayText: "81,200",
-          sellingPriceDisplayText: "40,600",
-        },
-        imageUrl: "pendant-lights.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "12:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 15116,
-          reviewAverage: 4.78,
-          reviewCountDisplayText: "15,116",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 70485,
-        },
-      },
-      {
-        id: "4",
-        name: "재생 우드 테이블",
-        brand: {
-          id: "4",
-          name: "에코플럭스",
-        },
-        price: {
-          originalPrice: "270800",
-          sellingPrice: "146200",
-          discountRate: "46",
-          originalPriceDisplayText: "270,800",
-          sellingPriceDisplayText: "146,200",
-        },
-        imageUrl: "wood-table.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 8% 쿠폰",
-            discountRatio: "8.0",
-            couponAppliedPrice: "23736",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 20126,
-          reviewAverage: 4.71,
-          reviewCountDisplayText: "20,126",
-          reviewAverageDisplayText: 4.7,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 191348,
-        },
-      },
-      {
-        id: "5",
-        name: "명상 의자",
-        brand: {
-          id: "5",
-          name: "젠플로우",
-        },
-        price: {
-          originalPrice: "170600",
-          sellingPrice: "112600",
-          discountRate: "34",
-          originalPriceDisplayText: "170,600",
-          sellingPriceDisplayText: "112,600",
-        },
-        imageUrl: "meditation-chair.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "13:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 5896,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "5,896",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 64555,
-        },
-      },
-      {
-        id: "6",
-        name: "달빛 무드등",
-        brand: {
-          id: "6",
-          name: "루나틱스",
-        },
-        price: {
-          originalPrice: "362000",
-          sellingPrice: "202700",
-          discountRate: "44",
-          originalPriceDisplayText: "362,000",
-          sellingPriceDisplayText: "202,700",
-        },
-        imageUrl: "mood-light.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 10% 쿠폰",
-            discountRatio: "10.0",
-            couponAppliedPrice: "35910",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 23967,
-          reviewAverage: 4.63,
-          reviewCountDisplayText: "23,967",
-          reviewAverageDisplayText: 4.6,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 167998,
-        },
-      },
-      {
-        id: "7",
-        name: "모자이크 커피 테이블",
-        brand: {
-          id: "7",
-          name: "픽셀퍼니",
-        },
-        price: {
-          originalPrice: "150600",
-          sellingPrice: "81300",
-          discountRate: "46",
-          originalPriceDisplayText: "150,600",
-          sellingPriceDisplayText: "81,300",
-        },
-        imageUrl: "coffee-table.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "12:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 5% 쿠폰",
-            discountRatio: "5.0",
-            couponAppliedPrice: "8455",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 9393,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "9,393",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 57065,
-        },
-      },
-      {
-        id: "8",
-        name: "부유 선반",
-        brand: {
-          id: "8",
-          name: "스페이스하모니",
-        },
-        price: {
-          originalPrice: "140400",
-          sellingPrice: "106700",
-          discountRate: "24",
-          originalPriceDisplayText: "140,400",
-          sellingPriceDisplayText: "106,700",
-        },
-        imageUrl: "shelf.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 10% 쿠폰",
-            discountRatio: "10.0",
-            couponAppliedPrice: "42210",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 16552,
-          reviewAverage: 4.64,
-          reviewCountDisplayText: "16,552",
-          reviewAverageDisplayText: 4.6,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 109802,
-        },
-      },
-      {
-        id: "9",
-        name: "천연 소재 러그",
-        brand: {
-          id: "9",
-          name: "오가닉스피릿",
-        },
-        price: {
-          originalPrice: "204700",
-          sellingPrice: "174000",
-          discountRate: "15",
-          originalPriceDisplayText: "204,700",
-          sellingPriceDisplayText: "174,000",
-        },
-        imageUrl: "rug.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "14:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 12428,
-          reviewAverage: 4.82,
-          reviewCountDisplayText: "12,428",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 30320,
-        },
-      },
-      {
-        id: "10",
-        name: "산업풍 침대 프레임",
-        brand: {
-          id: "10",
-          name: "메탈릭드림",
-        },
-        price: {
-          originalPrice: "428000",
-          sellingPrice: "273900",
-          discountRate: "36",
-          originalPriceDisplayText: "428,000",
-          sellingPriceDisplayText: "273,900",
-        },
-        imageUrl: "bed-frame.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "12:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 10% 쿠폰",
-            discountRatio: "10.0",
-            couponAppliedPrice: "3600",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 6937,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "6,937",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 74006,
-        },
-      },
-      {
-        id: "11",
-        name: "팝아트 벽 장식",
-        brand: {
-          id: "11",
-          name: "컬러버스트",
-        },
-        price: {
-          originalPrice: "331700",
-          sellingPrice: "218900",
-          discountRate: "34",
-          originalPriceDisplayText: "331,700",
-          sellingPriceDisplayText: "218,900",
-        },
-        imageUrl: "pop-art.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "14:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 11643,
-          reviewAverage: 4.73,
-          reviewCountDisplayText: "11,643",
-          reviewAverageDisplayText: 4.7,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 46988,
-        },
-      },
-      {
-        id: "12",
-        name: "공중 해먹",
-        brand: {
-          id: "12",
-          name: "에어리무브",
-        },
-        price: {
-          originalPrice: "200400",
-          sellingPrice: "134300",
-          discountRate: "33",
-          originalPriceDisplayText: "200,400",
-          sellingPriceDisplayText: "134,300",
-        },
-        imageUrl: "hammock.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "15:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 10% 쿠폰",
-            discountRatio: "10.0",
-            couponAppliedPrice: "11610",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 12887,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "12,887",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 80560,
-        },
-      },
-      {
-        id: "13",
-        name: "접이식 다이닝 테이블",
-        brand: {
-          id: "13",
-          name: "트랜스폼스페이스",
-        },
-        price: {
-          originalPrice: "230900",
-          sellingPrice: "184700",
-          discountRate: "20",
-          originalPriceDisplayText: "230,900",
-          sellingPriceDisplayText: "184,700",
-        },
-        imageUrl: "dining-table.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "15:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 10623,
-          reviewAverage: 4.76,
-          reviewCountDisplayText: "10,623",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 41473,
-        },
-      },
-      {
-        id: "14",
-        name: "우드 스톤 램프",
-        brand: {
-          id: "14",
-          name: "네이처블렌드",
-        },
-        price: {
-          originalPrice: "264600",
-          sellingPrice: "238200",
-          discountRate: "10",
-          originalPriceDisplayText: "264,600",
-          sellingPriceDisplayText: "238,200",
-        },
-        imageUrl: "wood-lamp.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "13:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 8945,
-          reviewAverage: 4.82,
-          reviewCountDisplayText: "8,945",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 48652,
-        },
-      },
-      {
-        id: "15",
-        name: "무소음 벽시계",
-        brand: {
-          id: "15",
-          name: "사일런트글로우",
-        },
-        price: {
-          originalPrice: "309300",
-          sellingPrice: "185600",
-          discountRate: "40",
-          originalPriceDisplayText: "309,300",
-          sellingPriceDisplayText: "185,600",
-        },
-        imageUrl: "wall-clock.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 9898,
-          reviewAverage: 4.66,
-          reviewCountDisplayText: "9,898",
-          reviewAverageDisplayText: 4.7,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 88306,
-        },
-      },
-      {
-        id: "16",
-        name: "부유형 침대",
-        brand: {
-          id: "16",
-          name: "플로팅드림",
-        },
-        price: {
-          originalPrice: "52000",
-          sellingPrice: "44700",
-          discountRate: "14",
-          originalPriceDisplayText: "52,000",
-          sellingPriceDisplayText: "44,700",
-        },
-        imageUrl: "bed.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 14995,
-          reviewAverage: 4.78,
-          reviewCountDisplayText: "14,995",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 115265,
-        },
-      },
-      {
-        id: "17",
-        name: "거울 벽 장식",
-        brand: {
-          id: "17",
-          name: "리플렉트아트",
-        },
-        price: {
-          originalPrice: "286900",
-          sellingPrice: "200800",
-          discountRate: "30",
-          originalPriceDisplayText: "286,900",
-          sellingPriceDisplayText: "200,800",
-        },
-        imageUrl: "mirror.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "14:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 8456,
-          reviewAverage: 4.73,
-          reviewCountDisplayText: "8,456",
-          reviewAverageDisplayText: 4.7,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 40475,
-        },
-      },
-      {
-        id: "18",
-        name: "안락 1인용 소파",
-        brand: {
-          id: "18",
-          name: "코지코브",
-        },
-        price: {
-          originalPrice: "193600",
-          sellingPrice: "133600",
-          discountRate: "31",
-          originalPriceDisplayText: "193,600",
-          sellingPriceDisplayText: "133,600",
-        },
-        imageUrl: "single-sofa.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: true,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "14:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 40924,
-          reviewAverage: 4.74,
-          reviewCountDisplayText: "40,924",
-          reviewAverageDisplayText: 4.7,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 85350,
-        },
-      },
-      {
-        id: "19",
-        name: "수저 세트",
-        brand: {
-          id: "19",
-          name: "아르코스",
-        },
-        price: {
-          originalPrice: "26500",
-          sellingPrice: "9900",
-          discountRate: "62",
-          originalPriceDisplayText: "26,500",
-          sellingPriceDisplayText: "9,900",
-        },
-        imageUrl: "cutlery-set.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "15:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-          couponBadge: {
-            displayText: "최대 5% 쿠폰",
-            discountRatio: "5.0",
-            couponAppliedPrice: "7505",
-          },
-        },
-        reviewStatistic: {
-          reviewCount: 7189,
-          reviewAverage: 4.77,
-          reviewCountDisplayText: "7,189",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 39844,
-        },
-      },
-      {
-        id: "20",
-        name: "폭신 폭신 베개",
-        brand: {
-          id: "20",
-          name: "아티젠",
-        },
-        price: {
-          originalPrice: "89400",
-          sellingPrice: "62600",
-          discountRate: "30",
-          originalPriceDisplayText: "89,400",
-          sellingPriceDisplayText: "62,600",
-        },
-        imageUrl: "pillow.png",
-        isSoldOut: false,
-        badgeProperties: {
-          isFreeDelivery: false,
-          isSpecialPrice: true,
-          isDiscontinued: false,
-          isSelling: true,
-          isOverseasPurchase: false,
-          isRetailDelivery: false,
-          isThirdPartyLogistic: false,
-          departureToday: {
-            orderDeadline: "15:00",
-            description: "평일 %s까지 결제시",
-            departureTodayExp: "B",
-          },
-          deliveryServiceCode: "NORMAL",
-        },
-        reviewStatistic: {
-          reviewCount: 8289,
-          reviewAverage: 4.77,
-          reviewCountDisplayText: "8,289",
-          reviewAverageDisplayText: 4.8,
-        },
-        scrapInfo: {
-          isScrap: false,
-          scrapCount: 42137,
-        },
-      },
-    ],
-    totalResults: 200,
+    goods: goods.slice(0, 20),
+    totalResults: 20,
   },
 };
