+++ src/shared/components/GoodItem/index.stories.tsx
@@ -35,31 +35,20 @@ export const Default: Story = {
         sellingPriceDisplayText: "154,000",
       },
       imageUrl: "cloud-sofa.png",
-      isSoldOut: false,
       badgeProperties: {
         isFreeDelivery: false,
         isSpecialPrice: true,
-        isDiscontinued: false,
-        isSelling: true,
-        isOverseasPurchase: false,
-        isRetailDelivery: false,
-        isThirdPartyLogistic: false,
         departureToday: {
           orderDeadline: "12:00",
           description: "평일 %s까지 결제시",
           departureTodayExp: "B",
         },
-        deliveryServiceCode: "NORMAL",
       },
       reviewStatistic: {
         reviewCount: 9405,
         reviewAverage: 4.79,
         reviewCountDisplayText: "9,405",
         reviewAverageDisplayText: 4.8,
-      },
-      scrapInfo: {
-        isScrap: false,
-        scrapCount: 87547,
       },
     },
   },
